creates spinner component
