+++ src/components/Spinner/index.tsx
@@ -1,0 +1,23 @@
+import React from 'react';
+import { HTMLAttributes } from 'react';
+
+interface IconProps extends HTMLAttributes<SVGElement> {}
+
+const Spinner: React.FC<IconProps> = (props) => (
+  <svg
+    xmlns="http://www.w3.org/2000/svg"
+    width="24"
+    height="24"
+    viewBox="0 0 24 24"
+    fill="none"
+    stroke="currentColor"
+    strokeWidth="2"
+    strokeLinecap="round"
+    strokeLinejoin="round"
+    {...props}
+  >
+    <path d="M21 12a9 9 0 1 1-6.219-8.56" />
+  </svg>
+);
+
+export default Spinner;
